+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## V1.5
+* [Feature] DNSLibs indicate that upstream uses DNSSEC. Turn on `ag::dnsproxy_settings::enable_dnssec_ok`
+    and check `ag::dns_request_processed_event::dnssec` in callback.
 * [Feature] Add ability to redirect all requests with DNS suffix to the fallbacks<p>
     see `ag::dnsproxy_settings::handle_dns_suffixes`
 * [Feature] DNS-over-QUIC default port changed. New port is 8853.
-    Now an address like `quic://dns.adguard.com` is transformed into `quic://dns.adguard.com:8853`
+    Now an address like `quic://dns.adguard.com` is transformed into `quic://dns.adguard.com:8853`.
     So to force the use of the old port `784` specify it strictly - `quic://dns.adguard.com:784`.
 * [Feature] Allow retrieving the library version
     * see `ag::dnsproxy::version()`